test: Test project organized
