+++ .github/workflows/sonarcloud.yml
@@ -65,7 +65,7 @@ jobs:
 
     - name: "Build and analyze"
       env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}  # Needed to get PR information, if any
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # Needed to get PR information, if any
         SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
         CollectCoverage: true # Hotfix for .NET 7
         CoverletOutputFormat: opencover # Hotfix for .NET 7
