+++ tests/PowerUtils.Results.Tests/ErrorTests/FailureErrorTests.cs
@@ -1,0 +1,87 @@
+﻿using System.Linq;
+using FluentAssertions;
+using PowerUtils.Results.Tests.Fakes;
+using Xunit;
+
+namespace PowerUtils.Results.Tests.ErrorTests
+{
+    public class FailureErrorTests
+    {
+        [Fact]
+        public void Error_Deconstruct_Properties()
+        {
+            // Arrange
+            var property = "fakeProperty";
+            var code = "fakeCode";
+            var description = "fakeDescription";
+
+
+            // Act
+            var (actProperty, actCode, actDescription) = Error.Failure(property, code, description);
+
+
+            // Assert
+            actProperty.Should().Be(property);
+            actCode.Should().Be(code);
+            actDescription.Should().Be(description);
+        }
+
+
+
+        [Fact]
+        public void Error_ImplicitResult_IsErrorTrue()
+        {
+            // Arrange
+            var property = "fakeFailureProperty";
+            var code = "fakeFailureCode";
+            var description = "fakeFailureDescription";
+
+
+            // Act
+            Result act = Error.Failure(property, code, description);
+
+
+            // Assert
+            act.IsError.Should()
+                .BeTrue();
+
+            act.Errors.Should()
+                .ContainSingle(s => s.Property == property);
+            act.Errors.Should()
+                .ContainSingle(s => s.Code == code);
+            act.Errors.Should()
+                .ContainSingle(s => s.Description == description);
+
+            act.Errors.First().Should().BeOfType<Error>();
+        }
+
+
+
+        [Fact]
+        public void Error_ImplicitValueResult_IsErrorTrue()
+        {
+            // Arrange
+            var property = "fakeFailureProperty";
+            var code = "fakeFailureCode";
+            var description = "fakeFailureDescription";
+
+
+            // Act
+            Result<FakeModel> act = Error.Failure(property, code, description);
+
+
+            // Assert
+            act.IsError.Should()
+                .BeTrue();
+
+            act.Errors.Should()
+                .ContainSingle(s => s.Property == property);
+            act.Errors.Should()
+                .ContainSingle(s => s.Code == code);
+            act.Errors.Should()
+                .ContainSingle(s => s.Description == description);
+
+            act.Errors.First().Should().BeOfType<Error>();
+        }
+    }
+}
