+++ tests/PowerUtils.Results.Tests/ErrorTests/GeneralErrorTests.cs
@@ -1,0 +1,51 @@
+﻿using FluentAssertions;
+using Xunit;
+
+namespace PowerUtils.Results.Tests.ErrorTests
+{
+    public class GeneralErrorTests
+    {
+        [Fact]
+        public void ErrorWithDescription_Construct_CustomDescription()
+        {
+            // Arrange
+            var property = "fakeProperty";
+            var code = "fakeCode";
+            var description = "fakeDescription";
+
+
+            // Act
+            var act = new Error(property, code, description);
+
+
+            // Assert
+            act.Property.Should()
+                .Be(property);
+            act.Code.Should()
+                .Be(code);
+            act.Description.Should()
+                .Be(description);
+        }
+
+        [Fact]
+        public void ErrorWithoutDescription_Construct_DefaultDescription()
+        {
+            // Arrange
+            var property = "fakeProperty";
+            var code = "fakeCode";
+
+
+            // Act
+            var act = new Error(property, code);
+
+
+            // Assert
+            act.Property.Should()
+                .Be(property);
+            act.Code.Should()
+                .Be(code);
+            act.Description.Should()
+                .Be($"An error has occurred with code '{code}'");
+        }
+    }
+}
