+++ tests/PowerUtils.Results.Tests/Fakes/FakeModel.cs
@@ -3,9 +3,16 @@
     public class FakeModel
     {
         public int Id { get; set; }
+
         public string Name { get; set; }
 
+
+
         public FakeModel() { }
+
+        public FakeModel(string name)
+
+            => Name = name;
         public FakeModel(int id, string name)
         {
             Id = id;
