+++ tests/PowerUtils.Results.Tests/VoidResultTests/ExplicitOperatorTests.cs
@@ -1,0 +1,40 @@
+﻿using FluentAssertions;
+using Xunit;
+
+namespace PowerUtils.Results.Tests.VoidResultTests
+{
+    public class ExplicitOperatorTests
+    {
+        [Fact]
+        public void ResultWithoutErrors_ImplicitAssignmentBool_ShouldBeFalse()
+        {
+            // Arrange
+            Result result = new Success();
+
+
+            // Act
+            var act = (bool)result;
+
+
+            // Assert
+            act.Should()
+                .BeTrue();
+        }
+
+        [Fact]
+        public void ResultWithErrors_ImplicitAssignmentBool_ShouldBeTrue()
+        {
+            // Arrange
+            Result result = Error.Forbidden("prop");
+
+
+            // Act
+            var act = (bool)result;
+
+
+            // Assert
+            act.Should()
+                .BeFalse();
+        }
+    }
+}
