+++ tests/PowerUtils.Results.Tests/VoidResultTests/MethodOkTests.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using FluentAssertions;
+using Xunit;
+
+namespace PowerUtils.Results.Tests.VoidResultTests
+{
+    public class MethodOkTests
+    {
+        [Fact]
+        public void WithoutErrors_GetErrors_InvalidOperationException()
+        {
+            // Arrange
+            var result = Result.Ok();
+
+
+            // Act
+            var act = Record.Exception(() => result.Errors);
+
+
+            // Assert
+            act.Should().BeOfType<InvalidOperationException>();
+            act.Message.Should().Be("Errors can be retrieved only when the result is an error");
+        }
+    }
+}
