+++ tests/PowerUtils.Results.Tests/VoidResultTests/MethodSuccessTests.cs
@@ -1,0 +1,20 @@
+﻿using FluentAssertions;
+using Xunit;
+
+namespace PowerUtils.Results.Tests.VoidResultTests
+{
+    public class MethodSuccessTests
+    {
+        [Fact]
+        public void SuccessMethod_CreateResult_IsErrorFalse()
+        {
+            // Arrange && Act
+            var act = Result.Success();
+
+
+            // Assert
+            act.IsError.Should()
+                .BeFalse();
+        }
+    }
+}
